% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{.read_dset_data}
\alias{.read_dset_data}
\title{.read_dset_data}
\usage{
.read_dset_data(group, name)
}
\arguments{
\item{group}{an open H5Group object.}

\item{name}{the name of a dataset in the group.}
}
\value{
a vector containing the data that was read.
}
\description{
.read_dset_data
}
