% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{read_h5_table}
\alias{read_h5_table}
\title{Read a tibble from a tabular HDF5 file.}
\usage{
read_h5_table(file, tablename)
}
\arguments{
\item{file}{an H5File object representing an HDF5 "tabular data" file, or the name of a file.}

\item{tablename}{The name of the table to be read.}
}
\value{
a tibble containing the data read from the named table.
}
\description{
Read a tibble from a tabular HDF5 file.
}
