% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{table_names}
\alias{table_names}
\title{Return the names of all tables in a tabular HDF5 file.}
\usage{
table_names(file)
}
\arguments{
\item{file}{an H5File object representing an HDF5 "tabular data" file, or the name of such a file.}
}
\value{
a vector (mode character) containing the names of the tables in the file.
}
\description{
Return the names of all tables in a tabular HDF5 file.
}
